service/requirements.txt
Flask==0.12.3
GitPython==2.1.8
